The python "only one way to do things" is broken
What I expect of a command line argument parser:
1. assign specific command line options to
appropriate functions and display the output
2. that's it

What do I get with dev-python/click:
1. decorators. Getting a proper two sentence
explanation on what they are is impossible,
so I just might as well write it here:

In python, functions are also objects, which
can be wrapped around. So when you "decorate"
a function you create another function that
calls the original function and replaces it
while doing some magic and passing values to
the original function.

Decorates are clones of your original function.
Only they are better, and now they can cook
spaghetti with some sauce added instead of
just cooking spaghetti.

2. An utility that preaches setuptools through
its documentations. There's no real-world
examples that do not use setuptools, at least,
I couldn't find any. The documentation has a
click-baity example that gives no clue on how
it will function in a real program.
Python was supposed to be a scripting language:
What the hell happened?
Why am I not allowed to create a single-file
python script with beautiful output? Why do
I need setuptools?

Forcing me to create what are effectively pip
modules is slave labor, and I'm not going to
do that. If I see this pattern emerging in
other languages I'm going back to C.

3. Adding an external dependency that doesn't do
the job.
Learning argparse vs. click + setuptools is probably
the same amount of hassle. With argparse, I only
depend on python. Guess who just won?

I'm totally in trouble at this point because
the portage API is written in python. And bridging
that to another language is hard in its own right.